+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic interface.
 - A cool logo.
 - A `cls` method to clear the command line interface.
+- `Player` class.